+++ src/policies/jwt-base.policy.ts
@@ -5,6 +5,7 @@ import { Policy } from '../decorators/policies/policy.decorator';
 import { JwtMetadata } from '../metadata/server/jwt';
 import { Logger } from '../services/logger';
 import { wrapPromise } from '../util/wrap-promise';
+import { HTTP_STATUS_UNAUTHORIZED } from '../util/http-status-type';
 
 type AbstractCtorT<T> = { (...args: any[]): T };
 
@@ -13,23 +14,35 @@ export class JwtBasePolicy {
     constructor(
         private jwtMeta: JwtMetadata,
         protected readonly logger: Logger,
-        credentialsRequired: boolean
+        public credentialsRequired: boolean
     ) {
         if (!jwtMeta) return;
         
         this.jwtHandler = expressJwt({
             secret: jwtMeta.secret,
             userProperty: this.property,
-            credentialsRequired: credentialsRequired
+            credentialsRequired: false
         });
     }
     
     get property() {
+        if (!this.jwtMeta) return undefined;
         return this.jwtMeta.tokenProperty;
     }
     private jwtHandler: RequestHandler;
     
     async handle(req: Request, res: Response) {
+        let jwt = await this.getJwt(req, res);
+        let reqProperty = this.property!;
+        if (jwt !== null) jwt = (<any>req)[reqProperty] = await this.fromJson(jwt);
+        if (this.credentialsRequired && !jwt) {
+            res.status(HTTP_STATUS_UNAUTHORIZED).json({ msg: 'Invalid Authorization header.' });
+        }
+        return jwt;
+    }
+    
+    private async getJwt(req: Request, res: Response) {
+        let reqProperty = this.property!
         if (this.jwtHandler) {
             try {
                 await wrapPromise<void>(this.jwtHandler, req, res);
@@ -37,11 +50,9 @@ export class JwtBasePolicy {
             catch (e) {
                 this.logger.verbose('jwt-policy', `express-jwt failed to parse 'Authorization' header.`);
                 this.logger.verbose('jwt-policy', `'Authorization' header: '${req.header('Authorization')}'`);
-                (<any>req)[this.property] = undefined;
             }
         }
-        if ((<any>req)[this.property]) return (<any>req)[this.property] = await this.fromJson((<any>req)[this.property]);
-        return null;
+        return (<any>req)[reqProperty] = (<any>req)[reqProperty] || null;
     }
     
     protected async fromJson(json: any) {